refactor: 모달 프레임 컴포넌트 리팩토링
