+++ README.md
@@ -174,7 +174,7 @@
   - [x] 모달 title 컴포넌트 분리
     - [x] 모달 title 컴포넌트 리팩토링
   - [x] 모달 프레임 컴포넌트 제작
-    - [x] 모달 프레임 컴포넌트 리팩토링
+    - [x] 모달 프레임 컴포넌트 리팩토링 (파일 이동 및 분리)
   - [x] 모달 백드롭 이벤트 등록
 - [ ] E2E 테스트 진행
 - [ ] 리팩토링
